finalized autonomous
this is the auton for skills, it works without the disks but when disks come into play it starts to not work. this is because the rollers make the intake suck in discs and mess up the movement. we need to figure out how to always keep the rollers running out until it needs to roll in to make sure the discs don't impact the movement.